Fix trouble to compile with MSVC
Remove std::forward here since MSVC does not compile with it. It issues a
cryptic error message:

...\duckdb-src\third_party\libpg_query\postgres_parser.cpp(43): error C2243: 'type cast': conversion from 'std::vector<duckdb_libpgquery::PGKeyword,std::allocator<duckdb_libpgquery::PGKeyword>> *' to '_Ty &&' exists, but is inaccessible
          with
          [
              _Ty=duckdb::vector<duckdb_libpgquery::PGKeyword,true>
          ]

But the real problem seems to be that std::forward should not really be used here.

Switch over to use std::move instead as suggested in https://github.com/duckdb/duckdb/discussions/12487
This does compile with MSVC, but issues a warning: C26479: Don't use std::move to return a local variable